+++ app/controllers/workouts_controller.rb
@@ -1,0 +1,22 @@
+class WorkoutsController < ApplicationController
+  def index
+    @week_start  = parse_week_start(params[:week_start])
+    @week_dates  = (0..6).map { |i| @week_start + i.days }
+    week_end     = @week_start + 6.days
+
+    plans = WorkoutPlan.where(planned_on: @week_start..week_end)
+    @workout_plans_by_date = plans.index_by(&:planned_on)
+
+    @week_counts = WorkoutPlan::TYPES.index_with do |type|
+      plans.count { |p| p.workout_type == type }
+    end
+  end
+
+  private
+
+  def parse_week_start(date_str)
+    Date.parse(date_str.to_s).beginning_of_week(:sunday)
+  rescue ArgumentError, TypeError
+    Time.zone.today.beginning_of_week(:sunday)
+  end
+end